+++ src/time_split/_backend/_splitter.py
@@ -4,7 +4,8 @@ from typing import cast, get_args
 
 from pandas import Timedelta, Timestamp
 from rics.misc import format_kwargs
-from rics.performance import format_seconds
+
+from time_split._compat import fmt_sec
 
 from ..settings import misc
 from ..types import (
@@ -12,7 +13,7 @@ from ..types import (
     DatetimeIterable,
     DatetimeSplitBounds,
     DatetimeSplits,
-    Flex,
+    ExpandLimits,
     Schedule,
     Span,
     TimedeltaTypes,
@@ -33,7 +34,7 @@ class DatetimeIndexSplitter:
     after: Span
     step: int
     n_splits: int
-    flex: Flex
+    expand_limits: ExpandLimits
 
     def get_splits(self, available: DatetimeIterable | None = None) -> DatetimeSplits:
         """Compute a split of given user data."""
@@ -48,7 +49,7 @@ class DatetimeIndexSplitter:
         return splits, ms
 
     def _materialize_schedule(self, available: DatetimeIterable | None = None) -> MaterializedSchedule:
-        ms = materialize_schedule(self.schedule, self.flex, available=available)
+        ms = materialize_schedule(self.schedule, self.expand_limits, available=available)
         if not ms.schedule.sort_values().equals(ms.schedule):
             raise ValueError(f"schedule must be sorted in ascending order; schedule={self.schedule!r} is not valid.")
 
@@ -137,10 +138,10 @@ class DatetimeIndexSplitter:
             if old == new:
                 return retval + "<no change>"
             diff = (new - old).total_seconds()
-            return retval + f"{_PrettyTimestamp(new).auto} ({'+' if diff > 0 else '-'}{format_seconds(abs(diff))})"
+            return retval + f"{_PrettyTimestamp(new).auto} ({'+' if diff > 0 else '-'}{fmt_sec(abs(diff))})"
 
         LOGGER.info(
-            f"Available data limits have been expanded (since flex={self.flex!r}):\n"
+            f"Available data limits have been expanded (since expand_limits={self.expand_limits!r}):\n"
             f"  start: {stringify(original[0], new=expanded[0])}\n"
             f"    end: {stringify(original[1], new=expanded[1])}"
         )